zod validation and show errors

+++ Database/src/products/products.controller.ts
@@ -10,9 +10,14 @@ import {
   BadRequestException,
 } from '@nestjs/common';
 import { ProductsService } from './products.service';
-import { CreateProductDto } from './productDto';
-import { UpdateProductDto } from './productDto';
+import {
+  CreateProductDto,
+  CreateProductValidator,
+  UpdateProductDto,
+  UpdateProductValidator,
+} from './productDto';
 import { Product } from '@prisma/client';
+import { fromZodError } from 'zod-validation-error';
 
 @Controller('products')
 export class ProductsController {
@@ -26,28 +31,29 @@ export class ProductsController {
   @Get(':id')
   async findOne(@Param('id') id: string): Promise<Product> {
     const idNumber = Number(id);
-    if (isNaN(idNumber)) {
+    if (isNaN(idNumber))
       throw new BadRequestException(
         'Invalid ID format. Please provide a valid number.',
       );
-    }
 
     const existedProduct = await this.productsService.findOne(idNumber);
-    if (existedProduct) {
-      return existedProduct;
-    }
+    if (existedProduct) return existedProduct;
 
-    throw new NotFoundException('Product not found');
+    throw new NotFoundException('Provided ID not found');
   }
 
   @Post()
   async create(@Body() createDto: CreateProductDto): Promise<Product> {
+    const parsedDto = CreateProductValidator.safeParse(createDto);
+    if (!parsedDto.success)
+      throw new BadRequestException(fromZodError(parsedDto.error).message);
+
     const existedProduct = await this.productsService.findByName(
       createDto.name,
     );
-    if (existedProduct) {
+    if (existedProduct)
       throw new BadRequestException('Provided name is already existed');
-    }
+
     return await this.productsService.create(createDto);
   }
 
@@ -56,20 +62,23 @@ export class ProductsController {
     @Param('id') id: string,
     @Body() updateDto: UpdateProductDto,
   ): Promise<Product> {
+    const parsedDto = UpdateProductValidator.safeParse(updateDto);
+    if (!parsedDto.success)
+      throw new BadRequestException(fromZodError(parsedDto.error).message);
+
     const idNumber = Number(id);
-    if (isNaN(idNumber)) {
+    if (isNaN(idNumber))
       throw new BadRequestException(
         'Invalid ID format. Please provide a valid number.',
       );
-    }
 
     const existedProduct = await this.productsService.findOne(idNumber);
     if (existedProduct) {
       existedProduct.updatedAt = new Date();
       return await this.productsService.update(idNumber, updateDto);
-    } else {
-      throw new NotFoundException('Product not found');
     }
+
+    throw new NotFoundException('Provided ID not found');
   }
 
   @Delete(':id')
@@ -82,10 +91,8 @@ export class ProductsController {
     }
 
     const existedProduct = await this.productsService.findOne(idNumber);
-    if (existedProduct) {
-      return this.productsService.remove(idNumber);
-    } else {
-      throw new NotFoundException('Product not found');
-    }
+    if (existedProduct) return this.productsService.remove(idNumber);
+
+    throw new NotFoundException('Provided ID not found');
   }
 }
